chore: Renames package serverpod_relic_helpers to relic.
