+++ packages/relic/lib/src/hijack_exception.dart
