+++ packages/relic/lib/src/middleware/add_chunked_encoding.dart
@@ -6,7 +6,7 @@ import 'dart:typed_data';
 
 import 'package:collection/collection.dart';
 import 'package:http_parser/http_parser.dart';
-import 'package:serverpod_relic_helpers/src/body.dart';
+import '../body.dart';
 
 import '../middleware.dart';
