+++ packages/relic/lib/src/util.dart
