+++ packages/relic/pubspec.yaml
@@ -1,4 +1,4 @@
-name: serverpod_relic_helpers
+name: relic
 description: A starting point for Dart libraries or applications.
 version: 1.0.0
 # repository: https://github.com/my_org/my_repo
