+++ packages/relic/test/add_chunked_encoding_test.dart
@@ -5,7 +5,7 @@
 import 'dart:async';
 import 'dart:typed_data';
 
-import 'package:serverpod_relic_helpers/serverpod_relic_helpers.dart';
+import 'package:relic/relic.dart';
 import 'package:test/test.dart';
 
 void main() {
