+++ packages/relic/test/message_test.dart
@@ -6,8 +6,8 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:typed_data';
 
-import 'package:serverpod_relic_helpers/src/body.dart';
-import 'package:serverpod_relic_helpers/src/message.dart';
+import 'package:relic/src/body.dart';
+import 'package:relic/src/message.dart';
 import 'package:test/test.dart';
 
 import 'test_util.dart';
