+++ packages/relic/test/response_test.dart
@@ -6,8 +6,7 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:typed_data';
 
-import 'package:serverpod_relic_helpers/serverpod_relic_helpers.dart'
-    hide Request;
+import 'package:relic/relic.dart' hide Request;
 import 'package:test/test.dart';
 
 import 'test_util.dart';
