+++ packages/relic/test/server_handler_test.dart
@@ -6,7 +6,7 @@
 
 import 'dart:async';
 
-import 'package:serverpod_relic_helpers/serverpod_relic_helpers.dart';
+import 'package:relic/relic.dart';
 import 'package:test/test.dart';
 
 import 'test_util.dart';
