+++ templates/pubspecs/packages/serverpod/pubspec.yaml
@@ -33,7 +33,7 @@ dependencies:
   system_resources: ^1.6.0
   vm_service: ">=13.0.0 <15.0.0"
   yaml: ^3.1.1
-  serverpod_relic_helpers: SERVERPOD_VERSION
+  relic: SERVERPOD_VERSION
   serverpod_shared: SERVERPOD_VERSION
   serverpod_serialization: SERVERPOD_VERSION
 
@@ -42,8 +42,8 @@ dev_dependencies:
   serverpod_lints: SERVERPOD_VERSION
 
 dependency_overrides:
-  serverpod_relic_helpers:
-    path: ../serverpod_relic_helpers
+  relic:
+    path: ../relic
   serverpod_shared:
     path: ../serverpod_shared
   serverpod_serialization:
