+++ tests/serverpod_test_server/pubspec.yaml
@@ -35,6 +35,8 @@ dev_dependencies:
   serverpod_test: 2.1.1
 
 dependency_overrides:
+  relic:
+    path: ../../packages/relic
   serverpod:
     path: ../../packages/serverpod
   serverpod_auth_server:
